backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
scikit-learn>=1.3.0
numpy>=1.20.0
pydantic==2.10.3
python-dotenv==1.0.1    
youtube-search-python==1.6.6
scipy>=1.12.0
pandas>=2.0.0
google-api-python-client==2.155.0
